requirements.txt
azure-functions
tweepy
requests
